backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
pywebpush
py-vapid
apscheduler